My  AI JARVIS/requirements.txt
# JARVIS Ultimate AI Assistant - Dependencies
# Install with: pip install -r requirements.txt

# Core speech
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13  # May need `pipwin install pyaudio` on Windows

# AI Providers (pick one or more)
groq>=0.4.0           # FREE - get key from console.groq.com
ollama>=0.1.0         # LOCAL - install from ollama.com
google-genai>=0.1.0   # Gemini - requires billing enabled

# System automation
pyautogui>=0.9.54
Pillow>=10.0.0

# System monitoring
psutil>=5.9.0

# Scheduling
schedule>=1.2.0

# Network utilities (optional)
speedtest-cli>=2.1.3

# Text processing
textblob>=0.17.1
